# (Micro?) optimize various models, especially accessors used to retrieve case values