--- a/v3/src/components/case-table/use-index-column.tsx
+++ b/v3/src/components/case-table/use-index-column.tsx
@@ -43,8 +43,8 @@ export const useIndexColumn = () => {
     const { __id__, [symIndex]: _index, [symParent]: parentId } = row
     const index = _index != null ? _index : data?.caseIndexFromID(__id__)
     const collapsedCases = (data && parentId && caseMetadata?.isCollapsed(parentId))
-                            ? data.pseudoCaseMap[parentId]?.childPseudoCaseIds?.length ??
-                              data.pseudoCaseMap[parentId]?.childCaseIds.length
+                            ? data.pseudoCaseMap.get(parentId)?.childPseudoCaseIds?.length ??
+                              data.pseudoCaseMap.get(parentId)?.childCaseIds.length
                             : undefined
     return (
       <IndexCell caseId={__id__} index={index} collapsedCases={collapsedCases} />