--- a/v3/src/models/data/data-set-collection-groups.test.ts
+++ b/v3/src/models/data/data-set-collection-groups.test.ts
@@ -1,3 +1,4 @@
+import { IAttribute } from "./attribute"
 import { CollectionModel } from "./collection"
 import { DataSet, IDataSet } from "./data-set"
 
@@ -71,7 +72,7 @@ describe("CollectionGroups", () => {
 
   it("handles grouping by a single attribute", () => {
     const collection = CollectionModel.create()
-    collection.addAttribute(data.attrFromID("aId"))
+    collection.addAttribute(data.attrFromID("aId") as IAttribute)
     data.addCollection(collection)
     expect(data.groupedAttributes.map(attr => attr.id)).toEqual(["aId"])
     expect(data.ungroupedAttributes.map(attr => attr.id)).toEqual(["bId", "cId"])
@@ -90,8 +91,8 @@ describe("CollectionGroups", () => {
 
   it("handles grouping by multiple attributes", () => {
     const collection = CollectionModel.create()
-    collection.addAttribute(data.attrFromID("aId"))
-    collection.addAttribute(data.attrFromID("bId"))
+    collection.addAttribute(data.attrFromID("aId") as IAttribute)
+    collection.addAttribute(data.attrFromID("bId") as IAttribute)
     data.addCollection(collection)
     expect(data.groupedAttributes.map(attr => attr.id)).toEqual(["aId", "bId"])
     expect(data.ungroupedAttributes.map(attr => attr.id)).toEqual(["cId"])
@@ -112,11 +113,11 @@ describe("CollectionGroups", () => {
 
   it("handles multiple groupings", () => {
     const collection1 = CollectionModel.create()
-    collection1.addAttribute(data.attrFromID("aId"))
+    collection1.addAttribute(data.attrFromID("aId") as IAttribute)
     data.addCollection(collection1)
     expect(data.collectionGroups.length).toBe(1)
     const collection2 = CollectionModel.create()
-    collection2.addAttribute(data.attrFromID("bId"))
+    collection2.addAttribute(data.attrFromID("bId") as IAttribute)
     data.addCollection(collection2)
     expect(data.groupedAttributes.map(attr => attr.id)).toEqual(["aId", "bId"])
     expect(data.ungroupedAttributes.map(attr => attr.id)).toEqual(["cId"])
@@ -219,7 +220,7 @@ describe("CollectionGroups", () => {
 
   it("doesn't take formula evaluated values into account when grouping", () => {
     const aAttr = data.attrFromID("aId")
-    aAttr.setDisplayExpression("foo * bar")
+    aAttr?.setDisplayExpression("foo * bar")
     data.moveAttributeToNewCollection("aId")
     expect(data.groupedAttributes.map(attr => attr.id)).toEqual(["aId"])
     expect(data.ungroupedAttributes.map(attr => attr.id)).toEqual(["bId", "cId"])