@pjanik 2023-12-01
Thanks @kswenson, I've merged or implemented all your suggestions.